Move not-really-CLI tests to misc
